API Update
I relocated all the display item code to the Core library. That stuff is
a bit complicated and easy to fuck up. I use it in a few plugins and it
can easily allow item duping if it's done wrong. This way I can manage
it all from one spot. I also added a couple of things for use a little
later.